% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aids2.R
\name{Aids2}
\alias{Aids2}
\title{Aids2 Dataset}
\format{
A data frame with 2843 observations and 7 variables:
\describe{
  \item{state}{State in Australia: NSW, Other, QLD, VIC}
  \item{sex}{Sex of the patient}
  \item{diag}{Date of diagnosis}
  \item{death}{Date of death}
  \item{status}{Status at the end of the study: A (alive), D (dead)}
  \item{T.categ}{Transmission category}
  \item{age}{Age at diagnosis}
}
}
\source{
Australian National AIDS Registry
}
\usage{
data(Aids2)
}
\description{
Australian AIDS survival data containing information on patients diagnosed with AIDS in Australia.
}
\examples{
data(Aids2)
head(Aids2)
summary(Aids2)
}
