% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdata.R
\name{bcdata}
\alias{bcdata}
\title{Boundary Condition Dataset}
\format{
A data frame with variables:
\describe{
  \item{time}{Time variable (seconds)}
  \item{pressure}{Pressure values (Pa)}
  \item{temperature}{Temperature values (K)}
  \item{velocity}{Velocity components (m/s)}
}
}
\source{
Experimental measurements or numerical simulations
}
\usage{
data(bcdata)
}
\description{
A dataset containing boundary condition data for computational fluid mechanics simulations.
}
\examples{
data(bcdata)
head(bcdata)
summary(bcdata)
}
