% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.CFM.R
\name{incremental_censored_factor_model}
\alias{incremental_censored_factor_model}
\title{Incremental Censored Factor Model}
\usage{
incremental_censored_factor_model(
  batch_data,
  m,
  max_iter = 100,
  tol = 1e-04,
  nugget = 1e-06
)
}
\arguments{
\item{batch_data}{List of data matrices (batches)}

\item{m}{Number of factors}

\item{max_iter}{Maximum number of ECM iterations (default: 100)}

\item{tol}{Convergence tolerance (default: 1e-4)}

\item{nugget}{Numerical stability term (default: 1e-6)}
}
\value{
A list containing model results
}
\description{
Implementation of incremental censored factor model for streaming data
with left-censored observations. Processes data in batches.
}
